Interactive mode (#165)
The changes introduced by this PR as follows:
* Adds `ie interactive` implementation to support executing markdown
documents interactively.
* Adds `ie inspect` to view the parsed data of a markdown document
produced by Innovation engines parser. This feature is still
experimental and only shows some of the parsing output for now.
* Updates the markdown parser to associate the last paragraph of
markdown document as the description for a codeblock.
* Adds a completely new rendering system using `bubble-tea` to support
interactive mode. Both `ie execute` and `ie test` will be rewritten in
the future to utilize a similar rendering implementation.
* Changes the status update for integration with the azure extensionto
nest codeblocks underneath steps and for codeblocks to include their
descriptions for learn mode.