Update the bash command execution to specify `set -e` when executing. (#155)
This PR allows for the innovation engine to correctly terminate a code
block after any command inside of it has failed. Previously the
innovation engine would rely on the last command in the code block being
executed to check for & report failure which meant that code blocks
which executed multiple commands would not fail if any command other
than the last one had an error.