fix(scenarios): Fix the regex that locates variables to override (#202)
This PR updates the regex used to find exported variables in a scenario
that have more complex declarations (I.E. variables that get values from
sub-shells, variables that are defined on the same line as another,
etc). This PR fixes #199 and tests have been added to ensure that this
issue doesn't arise again.