fix(scenario): Allow for markdown files to not specify h1 headers (#208)
h1 headers are currently used to denote the title of the scenario being
executed by IE and is currently made a requirement for the scenario to
be executed. This PR allows for markdown documents to not specify h1
tags by falling back to using the markdown file name as the title of the
scenario if no h1 tags are found within the document.